Move all tests into the test directory
