+++ tests/src/Functional/ThunderBaseTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Drupal\thunder;
+namespace Drupal\Tests\thunder\Functional;
 
 use Drupal\Tests\BrowserTestBase;
+use Drupal\Tests\thunder\Traits\ThunderTestTrait;
 
 /**
  * Class ThunderBaseTest.
